% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biometryPlot.R
\name{biometryPlot}
\alias{biometryPlot}
\title{Create Biometry Plot}
\usage{
biometryPlot(x, base_size, axis_text_size = 24, ...)
}
\arguments{
\item{x}{A data frame containing biometric measurements for a species.}

\item{base_size}{The base_size for the plot}

\item{axis_text_size}{The size of the axis text}

\item{...}{others ggplot parameters}
}
\value{
A biometry plot visualizing height, major diameter, and minor diameter.
}
\description{
This function generates a biometry plot based on the provided data frame.
It visualizes the measurements of height, major diameter, and minor diameter
for a specific species.
}
\details{
The function takes a data frame \code{x} containing the following columns:
\itemize{
\item \code{especie_code}: The species code.
\item \code{id_individuo}: The individual identifier.
\item \code{altura_cm}: Height in centimeters.
\item \code{dmayor_cm}: Major diameter in centimeters.
\item \code{dmenor_cm}: Minor diameter in centimeters.
}

The function creates a boxplot for each measurement type and uses custom colors.
It also includes half-eye plots and half-point plots.
}
