% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biometryStat.R
\name{biometryStat}
\alias{biometryStat}
\title{Calculate Biometry Statistics}
\usage{
biometryStat(x, variables = c("altura_cm", "dmayor_cm", "dmenor_cm"))
}
\arguments{
\item{x}{A data frame containing biometric data.}

\item{variables}{A character vector specifying the names of the biometric variables
for which statistics will be calculated. Default is c('altura_cm', 'dmayor_cm', 'dmenor_cm').}
}
\value{
A data frame with columns for each statistic (mean, sd, se, min, max) and a
corresponding variable column.
}
\description{
This function calculates various statistics (mean, standard deviation, standard error,
minimum, and maximum) for specified biometric variables in a data frame.
}
