% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityCommunity.R
\name{diversityCommunity}
\alias{diversityCommunity}
\title{Diversity Indices for Plant Communities}
\usage{
diversityCommunity(x)
}
\arguments{
\item{x}{A list containing data frames}
}
\value{
A data frame with diversity indices for plant communities.
}
\description{
This function computes diversity indices for plant communities based on the
specified method.
}
\details{
The function first extracts the method used for data collection from
the 'datos_generales' data frame. Based on the method, it selects the
corresponding plant community data frame and calculates various diversity indices.
}
