% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famexploreR-package.R
\docType{package}
\name{famexploreR-package}
\alias{famexploreR}
\alias{famexploreR-package}
\title{famexploreR: Exploration of Threathened Flora field data}
\description{
A shiny app that explores data from Threathened Flora. It uses customized forms to explore the data and generate several plots, tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ajpelu.github.io/famexploreR/}
}

}
\author{
\strong{Maintainer}: Antonio Jesús Pérez-Luque \email{ajpelu@gmail.com} (\href{https://orcid.org/0000-0002-1747-0469}{ORCID})

Authors:
\itemize{
  \item Juan Lorite \email{lorite@ugr} (\href{https://orcid.org/0000-0003-4617-8069}{ORCID}) [contributor]
}

}
\keyword{internal}
