% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herbivory.R
\name{herbivory}
\alias{herbivory}
\title{Calculate Herbivory Metrics and Create a Plot}
\usage{
herbivory(
  data,
  bar_color = "blue",
  point_fill = "green",
  point_color = "black",
  point_alpha = 0.9
)
}
\arguments{
\item{data}{A data frame containing the herbivory data with columns including
'comido_pct', 'id_individuo', and other relevant variables.}

\item{bar_color}{The color for the bar in the ggplot. Default is blue.}

\item{point_fill}{The fill color for points in the ggplot. Default is green.}

\item{point_color}{The outline color for points in the ggplot. Default is black.}

\item{point_alpha}{The alpha (transparency) for points in the ggplot. Default is 0.9.}
}
\value{
A list containing:
\itemize{
\item 'damage': A tibble with herbivory metrics.
\item 'plot_damage': A ggplot object displaying herbivory metrics.
}
}
\description{
This function calculates herbivory metrics from a given dataset and creates a
corresponding ggplot.
}
