% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborAbundance_stats.R
\name{neighborAbundance_stats}
\alias{neighborAbundance_stats}
\title{Calculate Neighbor Abundance Statistics}
\usage{
neighborAbundance_stats(data, units = c("m2", "dm2", "cm2"), focal_sp)
}
\arguments{
\item{data}{A data frame containing relevant columns, including "individuo"
(individual ID), "diam_muestreo_vecindad_cm" (neighborhood sampling diameter
in centimeters), "especie_vecina" (neighboring species), and "n_vecino"
(number of neighbors).}

\item{units}{A character vector specifying the units for area calculation.
Options are "m2" (square meters), "dm2" (square decimeters), and "cm2" (square
centimeters).}

\item{focal_sp}{A character string specifying the focal species.}
}
\value{
A list containing two data frames:
\describe{
\item{output}{A data frame with individual-level neighbor abundance statistics.}
\item{output_summary}{A data frame with summarized neighbor abundance statistics.}
}
}
\description{
This function calculates various neighbor abundance statistics based on input data.
}
