% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborSpecies_stats.R
\name{neighborSpecies_stats}
\alias{neighborSpecies_stats}
\title{Summarizes neighbor species data}
\usage{
neighborSpecies_stats(data)
}
\arguments{
\item{data}{A data frame or tibble containing the data to be summarized.}
}
\value{
A tibble with the following columns:
\itemize{
\item especie_vecina: The grouping variable (neighbor species).
\item ab_mean: The mean abundance of the neighbor species.
\item ab_se: The standard error of the mean abundance.
\item ab_min: The minimum abundance of the neighbor species.
\item ab_max: The maximum abundance of the neighbor species.
\item present_at: The count of plots where each neighbor species is recorded.
\item present_at_per: The percentage of the sampled plots (\code{individuo}) where each
neighbor species is recorded.
}
}
\description{
This function takes a data frame or tibble and groups it by neighbor species
(\code{especie_vecina}). For each neighbor species, the function calculates the
number of plots where each species has been recorded, as well as several
statistics for the neighbor species: the mean, standard error, minimum,
and maximum abundance of the neighbor species in the plots where it is present.
}
