% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCommunity.R
\name{plotCommunity}
\alias{plotCommunity}
\title{Plot Plant Community Data}
\usage{
plotCommunity(x, axis_text_size = 16, axis_title_size = 17, ...)
}
\arguments{
\item{x}{A list containing data frames for different aspects of the study.}

\item{axis_text_size}{The size of the axis text. Default value=16}

\item{axis_title_size}{The size of the axis title. Default value=17}

\item{...}{others ggplot parameters}
}
\value{
A bar plot visualizing plant community data.
}
\description{
This function generates a bar plot to visualize plant community data based on
the specified method.
}
\details{
The function first extracts the method used for data collection from
the 'datos_generales' data frame. Based on the method, it selects the
corresponding plant community data frame and creates a bar plot to visualize
the coverage of plant species.
}
