% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFlowering.R
\name{plotFlowering}
\alias{plotFlowering}
\title{Create a bar plot with error bars for flowering/fructification data.}
\usage{
plotFlowering(x, error = "se", bar_color = "blue", ...)
}
\arguments{
\item{x}{A data frame containing the flowering data.}

\item{error}{The type of error bars to use. Should be "se" (standard error) or
"sd" (standard deviation). Default is "se".}

\item{bar_color}{The color for the bars in the plot. Default is "blue".}

\item{...}{others ggplot parameters}
}
\value{
A ggplot2 object representing the bar plot with error bars.
}
\description{
This function generates a bar plot with error bars for flowering data, allowing
you to specify whether to use standard error ("se") or standard deviation ("sd")
error bars.
}
\seealso{
\code{\link{ggplot2}}, \code{\link{geom_bar}}, \code{\link{geom_errorbar}}
}
