% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareGeo.R
\name{prepareGeo}
\alias{prepareGeo}
\title{Prepare Geo Spatial Data}
\usage{
prepareGeo(x)
}
\arguments{
\item{x}{A data frame containing at least the columns 'campo' and 'valor',
where 'campo' specifies the type of information (e.g., 'crs', 'coord_x',
'coord_y', 'elevacion'), and 'valor' contains the corresponding values.}
}
\value{
A Simple Features (sf) object with spatial coordinates and attributes.
}
\description{
This function takes a data frame containing information about geographic
coordinates and prepares it for spatial analysis by converting it into a
Simple Features (sf) object.
}
\details{
This function creates an sf object with spatial coordinates and
attributes, from an input data frame thata contain specific columns
for 'campo' and 'valor'.
}
