% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparePopup.R
\name{preparePopup}
\alias{preparePopup}
\title{preparaPopup function}
\usage{
preparePopup(
  x,
  mdfields = c("especie focal", "localidad", "site", "reference", "poblacion",
    "tratamiento", "elevacion", "fecha")
)
}
\arguments{
\item{x}{A tibble containing data with columns 'campo' and 'valor'.}

\item{mdfields}{A vector of field names specifying which fields to include in
the popup content.}
}
\value{
A character string containing HTML-formatted popup content.
}
\description{
This function takes a tibble \code{x} and a vector of field names \code{mdfields} as
input and prepares an HTML popup content based on the specified fields.
}
