% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAllsheets.R
\name{readAllsheets}
\alias{readAllsheets}
\title{Read data from all sheets of an uploaded file.}
\usage{
readAllsheets(upload_path, valid_sheets)
}
\arguments{
\item{upload_path}{Character string specifying the path to the uploaded file.}

\item{valid_sheets}{Character vector specifying the names of sheets that are
expected to be in the uploaded file.}
}
\value{
A named list containing data from all sheets in the uploaded file,
with sheet names as list names.
}
\description{
This function reads data from all sheets of an uploaded file in ODS or XLSX
format and returns them in a named list.
}
