% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternaryPlot.R
\name{ternaryPlot}
\alias{ternaryPlot}
\title{Generate a Ternary Plot with Customizable Axis Labels}
\usage{
ternaryPlot(data, xvar, yvar, zvar, bsize, point_size, ...)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{xvar}{The name of the variable to be plotted in the x-axis of the ternary plot.}

\item{yvar}{The name of the variable to be plotted in the y-axis of the ternary plot.}

\item{zvar}{The name of the variable to be plotted in the z-axis of the ternary plot.}

\item{bsize}{The base font size for the plot.}

\item{point_size}{The size of the points.}
}
\value{
A ggtern plot object displaying the ternary plot.
}
\description{
This function generates a ternary plot using the ggtern package with customizable
axis labels. The function allows specifying the variable names for the x, y,
and z axes. It also capitalizes the first letter of each variable name for axis labels.
}
\details{
This function creates a ternary plot, where data points are represented
by points in a triangular coordinate system. The \code{xvar}, \code{yvar}, and \code{zvar}
arguments allow you to specify which variables from the \code{data} argument should
be used for each axis. The function also capitalizes the first letter of each
variable name for use in the axis labels. You can adjust the base font size
(\code{bsize}) for the plot to control the text size.
}
\examples{
\dontrun{
# Example usage of the custom function with customized variable names
data <- data.frame(
  arena = c(0.4, 0.3, 0.2, 0.1),
  arcilla = c(0.3, 0.4, 0.2, 0.1),
  limo = c(0.3, 0.3, 0.4, 0.1)
)

generateTernaryPlot(data, xvar = 'arena', yvar = 'arcilla', 
zvar = 'limo')
}


}
\seealso{
\code{\link{ggtern}} for more information on creating ternary plots using ggtern.
}
