% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecindadPlot.R
\name{vecindadPlot}
\alias{vecindadPlot}
\title{Generate Vecindad Plot}
\usage{
vecindadPlot(x, ...)
}
\arguments{
\item{x}{A data frame containing data for plotting.
\itemize{
\item \code{especie_vecina}: The neighboring species.
\item \code{ab_mean}: The mean abundance of the neighboring species.
\item \code{ab_se}: The standard error of the mean abundance.
}}
}
\value{
A bar plot with error bars.
}
\description{
This function generates a bar plot with error bars that visualizes the abundance
of neighboring species.
}
